WeaponCardGenerator Script Updated
"As a gamedev, I want the weapon cards to be procedurally generated so that I do not have to make individual cards for each weapon"

Text positioning and stylisation updated. Weapon Dictionary entries complete for Sword, Hammer and Cutter. Dictionary will need to be updated as more weapons are added.

Weapon Description complete for Sword. Still needs to be updated for Hammer, Cutter and any other weapons added in the future